+++ crates/printer/src/truncation.rs
@@ -3,8 +3,6 @@ use std::slice::IterMut;
 
 use types::FilteredItem;
 
-pub const DOTS: &str = "..";
-
 /// Line number of Vim is 1-based.
 pub type VimLineNumber = usize;
 
@@ -26,79 +24,33 @@ const WINWIDTH_OFFSET: usize = 4;
 #[cfg(test)]
 const WINWIDTH_OFFSET: usize = 0;
 
-// https://stackoverflow.com/questions/51982999/slice-a-string-containing-unicode-chars
-#[inline]
-pub fn utf8_str_slice(line: &str, start: usize, end: usize) -> String {
-    line.chars().take(end).skip(start).collect()
-}
-
-fn truncate_line_impl(
+fn truncate_line_v1(
     line: &str,
-    indices: &[usize],
+    indices: &mut [usize],
     winwidth: usize,
     skipped: Option<usize>,
 ) -> Option<(String, Vec<usize>)> {
-    let last_idx = indices.last()?;
-    if *last_idx > winwidth {
-        let mut start = *last_idx - winwidth;
-        if start >= indices[0] || (indices.len() > 1 && *last_idx - start > winwidth) {
-            start = indices[0];
-        }
-        let line_len = line.len();
-        // [--------------------------]
-        // [-----------------------------------------------------------------xx--x--]
-        for _ in 0..3 {
-            if indices[0] - start >= DOTS.len() && line_len - start >= winwidth {
-                start += DOTS.len();
-            } else {
-                break;
-            }
-        }
-        let trailing_dist = line_len - last_idx;
-        if trailing_dist < indices[0] - start {
-            start += trailing_dist;
-        }
-        let end = line.len();
-        let left_truncated = if let Some(n) = skipped {
-            let icon: String = line.chars().take(n).collect();
-            format!("{}{}{}", icon, DOTS, utf8_str_slice(line, start, end))
-        } else {
-            format!("{}{}", DOTS, utf8_str_slice(line, start, end))
-        };
+    use crate::trimmer::v1::trim_text;
 
-        let offset = line_len.saturating_sub(left_truncated.len());
+    if line.is_empty() || indices.is_empty() {
+        return None;
+    }
 
-        let left_truncated_len = left_truncated.len();
-
-        let (truncated, max_index) = if left_truncated_len > winwidth {
-            if left_truncated_len == winwidth + 1 {
-                (
-                    format!("{}.", utf8_str_slice(&left_truncated, 0, winwidth - 1)),
-                    winwidth - 1,
-                )
-            } else {
-                (
-                    format!(
-                        "{}{}",
-                        utf8_str_slice(&left_truncated, 0, winwidth - 2),
-                        DOTS
-                    ),
-                    winwidth - 2,
-                )
-            }
-        } else {
-            (left_truncated, winwidth)
-        };
-
-        let truncated_indices = indices
-            .iter()
-            .map(|x| x - offset)
-            .take_while(|x| *x < max_index)
-            .collect::<Vec<_>>();
-
-        Some((truncated, truncated_indices))
+    if let Some(skipped) = skipped {
+        let container_width = winwidth - skipped;
+        let text = line.chars().skip(skipped).collect::<String>();
+        indices.iter_mut().for_each(|x| *x -= 2);
+        trim_text(&text, indices, container_width, 4).map(|(text, mut indices)| {
+            (
+                format!("{}{}", line.chars().take(skipped).collect::<String>(), text),
+                {
+                    indices.iter_mut().for_each(|x| *x += 2);
+                    indices
+                },
+            )
+        })
     } else {
-        None
+        trim_text(line, indices, winwidth, 4)
     }
 }
 
@@ -115,11 +67,35 @@ pub fn truncate_long_matched_lines<T>(
 ) -> LinesTruncatedMap {
     let mut truncated_map = HashMap::new();
     let winwidth = winwidth - WINWIDTH_OFFSET;
+    items.enumerate().for_each(|(lnum, mut filtered_item)| {
+        let source_item = &filtered_item.source_item;
+        if let Some((truncated, truncated_indices)) = truncate_line_v1(
+            source_item.display_text(),
+            &mut filtered_item.match_indices,
+            winwidth,
+            skipped,
+        ) {
+            truncated_map.insert(lnum + 1, source_item.display_text().to_string());
+
+            filtered_item.display_text = Some(truncated);
+            filtered_item.match_indices = truncated_indices;
+        }
+    });
+    truncated_map
+}
+
+pub fn truncate_long_matched_lines_v0<T>(
+    items: IterMut<FilteredItem<T>>,
+    winwidth: usize,
+    skipped: Option<usize>,
+) -> LinesTruncatedMap {
+    let mut truncated_map = HashMap::new();
+    let winwidth = winwidth - WINWIDTH_OFFSET;
     items.enumerate().for_each(|(lnum, filtered_item)| {
         let line = filtered_item.source_item_display_text();
 
         if let Some((truncated, truncated_indices)) =
-            truncate_line_impl(line, &filtered_item.match_indices, winwidth, skipped)
+            crate::trimmer::v0::trim_text(line, &filtered_item.match_indices, winwidth, skipped)
         {
             truncated_map.insert(lnum + 1, line.to_string());
 
@@ -142,11 +118,11 @@ pub fn truncate_grep_lines(
     let (lines, indices): (Vec<String>, Vec<Vec<usize>>) = lines
         .into_iter()
         .zip(indices.into_iter())
-        .map(|(line, indices)| {
+        .map(|(line, mut indices)| {
             lnum += 1;
 
             if let Some((truncated, truncated_indices)) =
-                truncate_line_impl(&line, &indices, winwidth, skipped)
+                truncate_line_v1(&line, &mut indices, winwidth, skipped)
             {
                 truncated_map.insert(lnum, line);
                 (truncated, truncated_indices)